Add iterator support for collections
To help users page through collections, this adds an option to the
SDK to return iterators from the default response handler for
collection responses (i.e. GET requests that return an object with
the entries field).  The iterator conforms to the proposed spec
for async iterators, which is currently nearing the final stages
for being accepted into a future ES version. See
https://github.com/tc39/proposal-async-iteration for more details.

This approach maximizes future compatibility (e.g. with upcoming
syntax like for-await-of loops), while minimizing the surface area
of the change and keeping it backwards-compatible.