Support passing client IP to all token methods (#155)
Add support for passing an optional `ip` parameter to all token methods via the SDK interface, as well as for automatically passing a client's IP through to all token operations made by the client.  This ensures that a client with an IP address set will always report that IP correctly to Box.