Fix token epxiration window
Fix the 150-second window between token staleness and expiration
where two Persistent Clients sharing a TokenStore could cause
one of the clients to generate 401 errors for all calls.  This
change removes the distinction between "stale" and "expired" tokens
and deprecates the staleBufferMS config option in favor of a single
expiredBufferMS option.  The old option will still be honored, but
tokens will be considered expired at the maximum of the stale
and expired times.