Bump jsonwebtoken version
We were previously relying on a loose semver range to provide the
correct version of the jwa library (we need at least v1.1.0).  The
correct version was in our semver range, but we would also accept
lower versions that gave buggy behavior.  This commit bumps us to
the latest version of jsonwebtoken, which itself requires the right
version of the lower-level jws and jwa libraries.