Allow reconfiguring of request retry strategy (#337)
Adds `config.retryStrategy` which is an optional config param. It is a function that takes in an object of options including the error, # of retries made so far, and the `config.numMaxRetries` and `config.retryIntervalMS`.  If the function returns a number, then this will be used as the time (in ms) until the next retry. If the function returns an Error, this will be propagated in the response to the user. If neither of those are returned, the original response error object will be propagated.